requirements.txt
google-adk>=1.0.0
requests>=2.31.0
pymupdf>=1.24.0
arxiv>=2.1.0
